+++ src/Model/Callback/TakeThreadControl.php
@@ -40,6 +40,6 @@ class TakeThreadControl
      */
     public static function create(array $callbackData): self
     {
-        return new self($callbackData['previous_owner_app_id'], $callbackData['metadata']);
+        return new self($callbackData['previous_owner_app_id'], $callbackData['metadata'] ?? null);
     }
 }