+++ tests/TestCase/Api/ProfileTest.php
@@ -62,7 +62,7 @@ class ProfileTest extends AbstractTestCase
         $this->assertSame('success', $response->getResult());
     }
 
-    public function testBadField(): void
+    public function testInvalidField(): void
     {
         $this->expectException(MessengerException::class);
         $this->expectExceptionMessage('menu is not a valid value. fields must only contain "get_started, greeting, ice_breakers, persistent_menu, whitelisted_domains, account_linking_url, payment_settings, home_url, target_audience".');