+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The Messenger library follows [SemVer](http://semver.org/).
 
 ## 3.x
 
+**Changelog** (since [`3.3.1`](https://github.com/ker0x/messenger/compare/3.3.1...3.3.2))
+
+- 3.3.2 (2020-02)
+    - Allow to pass an array as recipient to send message
+    - Rename all `testBad*` methods to `testInvalid*` 
+
 **Changelog** (since [`3.3.0`](https://github.com/ker0x/messenger/compare/3.3.0...3.3.1))
 
 - 3.3.1 (2020-01)